+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes will be documented in this file.
 
+## 1.0.1 - 2022-01-18
+
+- Fixed phpstan (1.4) level 6
+- After install example.config.php is copied to config/relax.php
+
 ## 1.0.0 - 2022-01-18
 
 - initial release